requirements.txt
asgiref==3.4.1
Django==3.2.8
django-classy-tags==2.0.0
django-cms==3.9.0
django-filer==2.0.2
django-formtools==2.3
django-js-asset==1.2.2
django-mptt==0.13.4
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-snippet==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
easy-thumbnails==2.7.2
html5lib==1.1
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
Unidecode==1.1.2
webencodings==0.5.1

